+++ src/jupyterscad/_render.py
@@ -14,21 +14,14 @@ PARTICULAR PURPOSE. See the GNU General Public License for more details.
 You should have received a copy of the GNU General Public License along with
 this program. If not, see <https://www.gnu.org/licenses/>.
 """
-import logging
-import platform
-import subprocess
 import tempfile
-from pathlib import Path
+from os import PathLike
+from typing import Optional, Union
 
+import pythreejs as pjs
+
+from ._openscad import process
 from ._visualize import visualize_stl
-from .exceptions import OpenSCADException
-
-DEFAULT_OPENSCAD_EXECUTABLE = {
-    "Linux": ["/usr/bin/openscad", "/usr/local/bin/openscad"],
-    "Darwin": ["/Applications/OpenSCAD.app/Contents/MacOS/OpenSCAD"],
-}
-
-LOGGER = logging.getLogger(__name__)
 
 
 def render(
@@ -36,16 +29,14 @@ def render(
     width: int = 400,
     height: int = 400,
     grid_unit: float = 1,
-    outfile: str = None,
-    openscad_exec: Path = None,
-):
+    outfile: Optional[Union[str, PathLike]] = None,
+    openscad_exec: Optional[Union[str, PathLike]] = None,
+) -> pjs.Renderer:
     """Render a visusualization of an OpenSCAD object.
 
     Typical usage example:
 
-      obj = cube(3)
-      r = render(obj)
-      display(r)
+        >>> display(render(cube(3)))
 
     Args:
         obj: OpenSCAD object to visualize.
@@ -59,27 +50,30 @@ def render(
         Rendering to be displayed.
 
     Raises:
-        OpenSCADException: An error occurred running OpenSCAD.
+        exceptions.OpenSCADException: An error occurred running OpenSCAD.
     """
     if outfile:
-        render_stl(str(obj), outfile, openscad_exec=openscad_exec)
+        render_stl(obj, outfile, openscad_exec=openscad_exec)
         r = visualize_stl(outfile, width=width, height=height, grid_unit=grid_unit)
     else:
         with tempfile.NamedTemporaryFile(suffix=".stl", delete=False) as stl_tmp_file:
-            render_stl(str(obj), stl_tmp_file.name, openscad_exec=openscad_exec)
+            render_stl(obj, stl_tmp_file.name, openscad_exec=openscad_exec)
             r = visualize_stl(
                 stl_tmp_file.name, width=width, height=height, grid_unit=grid_unit
             )
     return r
 
 
-def render_stl(obj, output_file, openscad_exec: Path = None):
+def render_stl(
+    obj,
+    outfile: Union[str, PathLike],
+    openscad_exec: Optional[Union[str, PathLike]] = None,
+):
     """Render a stl from an OpenSCAD object.
 
     Typical usage example:
 
-      obj = cube(3)
-      render_stl(obj, outfile='cube.stl')
+        >>> render_stl(cube(3), 'cube.stl')
 
     Args:
         obj: OpenSCAD object to visualize.
@@ -87,69 +81,10 @@ def render_stl(obj, output_file, openscad_exec: Path = None):
         openscad_exec: Path to openscad executable.
 
     Raises:
-        OpenSCADException: An error occurred running OpenSCAD.
+        exceptions.OpenSCADException: An error occurred running OpenSCAD.
     """
     with tempfile.NamedTemporaryFile(suffix=".scad", delete=False) as scad_tmp_file:
         with open(scad_tmp_file.name, "w") as fp:
             fp.write(str(obj))
 
-        OpenSCAD(openscad_exec).render(scad_tmp_file.name, output_file)
-
-
-class OpenSCAD:
-    def __init__(self, openscad_exec: Path = None):
-        if openscad_exec:
-            if not self._is_executable(openscad_exec):
-                raise OpenSCADException(
-                    f"Provided path to openscad executable ({openscad_exec}) does not "
-                    "exist."
-                )
-            self.executable = openscad_exec
-        else:
-            self.executable = self._detect_executable()
-
-    @classmethod
-    def _detect_executable(cls) -> Path:
-        detected_executable = None
-        system = platform.system()
-        try:
-            default_paths = DEFAULT_OPENSCAD_EXECUTABLE[system]
-        except KeyError:  # system not supported
-            raise OpenSCADException(
-                f"This system ({system}) is not supported for "
-                "OpenSCAD executable autodetect. Please specify the path "
-                "to the OpenSCAD executable."
-            )
-        else:
-            for test_path in default_paths:
-                if cls._is_executable(test_path):
-                    LOGGER.debug(
-                        f"executable path ({test_path}) found for "
-                        f"system ({system})."
-                    )
-                    detected_executable = Path(test_path)
-                    break
-
-            if not detected_executable:
-                raise OpenSCADException(
-                    "OpenSCAD executable not found at the default "
-                    f"locations for this system ({system}), "
-                    f"{DEFAULT_OPENSCAD_EXECUTABLE[system]}. "
-                    "Please specify the path to the OpenSCAD "
-                    "executable."
-                )
-
-        return detected_executable
-
-    @staticmethod
-    def _is_executable(path: str):
-        return Path(path).is_file()
-
-    def render(self, scad_file, output_file):
-        """Generate stl from scad"""
-        cmd = [self.executable, "-o", output_file, scad_file]
-        LOGGER.info(cmd)
-        try:
-            subprocess.run(cmd, check=True, capture_output=True, text=True)
-        except subprocess.CalledProcessError as e:
-            raise OpenSCADException(str(e.stderr))
+        process(scad_tmp_file.name, outfile, executable=openscad_exec)