+++ src/jupyterscad/exceptions.py
@@ -14,11 +14,22 @@ PARTICULAR PURPOSE. See the GNU General Public License for more details.
 You should have received a copy of the GNU General Public License along with
 this program. If not, see <https://www.gnu.org/licenses/>.
 """
+import sys
 
 
-class JupyterSCADException(Exception):
+class JupyterSCADError(Exception):
     pass
 
 
-class OpenSCADException(JupyterSCADException):
+class OpenSCADError(JupyterSCADError):
     pass
+
+
+class RenderError(JupyterSCADError):
+    def __init__(self, message: str, src: str) -> None:
+        super().__init__(message)
+        self.message = message
+        self.src = src
+
+    def show(self) -> None:
+        print(f"{self.message}\nSCAD SOURCE:\n\n{self.src}", file=sys.stderr)