+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - API documentation
 
+### Fixed
+
+- README references to jupter_scad should be jupyterscad
+
 ## [0.1.0] - 2023-10-01
 
 ### Added