+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - OpenSCAD executable discovery now looks in PATH (which works for Linux) then
   in the macOS-specific install path if not found in PATH
+- Rename render_stl out_file to outfile for positional arg to match other fcns
 
 ## [0.1.0] - 2023-10-01